requirements.txt
fastapi
uvicorn
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
python-dotenv
